Replace usage of socketpair for IPC
On Windows, using a socket as standard input file descriptor is
daring. Even if it may be possible, Lwt will currently outright reject
it. We can, instead of passing a file descriptor to the spawned
process, pass the address of the UNIX domain socket as an argument so
that the sub process may open a socket itself and connect to that
address.

UNIX domain sockets are supported on Windows 10 and OCaml 4.14.