Replace usage of socketpair for IPC
Everything works as expected on Linux, but the following sections
apply on Windows.

Using a socket as standard input file descriptor is daring. Even if it
may be possible, Lwt will currently outright reject it. We can,
instead of passing a file descriptor to the spawned process, pass the
address of the UNIX domain socket as an argument so that the sub
process may open a socket itself and connect to that address.

Creating a filesystem entry with the same name as the socket seems
also to be a problem. Testing that the filesystem entry for the socket
exists with `Sys.file_exists` (`_wstati64`), or `Cmdliner.Arg.file` is
currently buggy (possibly related to [9793]); but we can still unlink
the socket.

[9793]: https://github.com/ocaml/ocaml/pull/9793

Although Unix domain sockets might appear in the filesystem, it's all
a lie made by the kernel. Windows currently doesn't support abstract
socket, or unnamed sockets returned by `socketpair` (`socketpair`
isn't implemented).

UNIX domain sockets are supported on Windows 10 and OCaml 4.14.

Give 1 second for the solver process to connect to the socket. If not,
or if the parent gets a signal that the child has exited, fail. Keep
the failure code simple because we're not in Lwt loop yet, and because
this shouldn't happen.