oldest_commits_with: fix caching
The current version has an important issue: `oldest_commit_with` always
returns the commit equals to `from`. In other words, the most recent
commit of `opam-repository`. This is correct but destroys caching
because every `opam-repository` push will trigger a rebuild of all
ocaml-ci repositories.

The reason is that the `paths` argument to that function is `[]`. In
turn, this is caused by `is_path_in_repo` always returning `false`.
There are two issues in that function:
- the string passed as a second argument to `/` is not a valid segment,
  so it raises
- the clone is a bare one, so it does not contain any files

Removing this filtering makes the function return an older commit (and
it is possible to observe that paths are being passed to `git log`, for
example using `execsnoop`).