Fix bug: 'Bad frame from worker' and  'Channel_closed'
This bug is caused when the pool reuse an internal-worker(worker process) which
is going to be terminated because a kill signal is already sent to the
internal-worker due to a job cancelling. The operating system take some time to
kill a process when a kill signal is sent.

The 'Bad frame from worker' is when the controller try to read from that
internal-worker(released and being terminated), in the previous use some part of
internal-worker output is read.

There's also a current fail because of Lwt_io.Channel_closed. the
the controller also start reading when the internal-worker is killed by the OS
and all the channels are closed.