Add a cache system
The response of a request is cached and it is invalidated every time one
of then packages of the new request is involve ind the differencies
between the new opam commit and the old one(cached opam commit).

Anytime the cached response is the same as the new response, we keep the
old one for the oldest opam commits which are used.