Use OCaml code to find the oldest commit
Before, we spawned a separate `git log` process after each solve.
This has the advantage that it can do work independently of the OCaml
GC, but it also has some disadvantages:

- It has to read and decompress the repository for each solve, rather
  than caching it. This is slow.
- Spawning a process requires briefly pausing all domains, slowing down
  solving.
- The spawned processes might run on a CPU that OCaml wants to use,
  delaying that domain and, in turn, disrupting the STW minor GC.

On my 8-core machine, I get this (running `bench.sh`):

    Workers	Before	After
    1	        2.79	2.99
    2	        4.93	5.4
    3	        6.76	7.76
    4	        8.37	9.47
    5	        9.77	11.3
    6	        10.27	12.49
    7	        11.07	14.09
    8	        7.73	8.95

Just looking at the time to find the commit, I get:

    Before After
    0.035s 0.246s
    0.186s 1.727s
    0.039s 0.000s
    0.038s 0.000s
    0.038s 0.000s
    0.037s 0.001s
    0.039s 0.000s
    0.040s 0.000s
    0.040s 0.001s
    ...

That is, the OCaml code is a lot slower at loading the data from disk,
but once cached it's massively faster.